♻️ refactor theme/color set
