+++ src/theme/index.js
@@ -1,0 +1,10 @@
+const set = require("./set");
+// next features...
+// get
+// find
+
+module.exports = {
+    set
+    //get
+    //find
+}
